necromancer almost done

+++ Assets/Scripts/Actions/Explode.cs
@@ -11,7 +11,7 @@ public partial class Explode : Action
 {
     [SerializeReference] public BlackboardVariable<GameObject> Agent;
     [SerializeReference] public BlackboardVariable<ParticleSystem> Particle;
-    [SerializeReference] public BlackboardVariable<float> Damage;
+    [SerializeReference] public BlackboardVariable<int> Damage;
     [SerializeReference] public BlackboardVariable<float> Radius;
     [SerializeReference] public BlackboardVariable<float> Force;
 
@@ -34,7 +34,7 @@ public partial class Explode : Action
 
         Vector3 explosionCenter = Agent.Value.transform.position;
         float expRadius = Radius.Value;
-        float expDamage = Damage.Value;
+        int expDamage = Damage.Value;
         float expForce = Force.Value;
